ml: the event frame casts mapped p90, so a categorical variable can divide
The 2017 district retrain trained all eight regressors and then died after
10,425 s in build_event_frame:

    df["conf"] = (1.0 - df["pred_err"] / df["p90"]).clip(...)
    TypeError: Object with dtype category cannot perform the numpy op divide

a2f0e84 made the paired frame's `variable` categorical to fit a year in
memory. pandas' Series.map on a categorical returns a categorical whenever the
mapping is one-to-one, and arithmetic on it raises. At full scale every
variable's p90 differs, so the map is one-to-one; the existing test mapped every
variable to the same 5.0, which pandas returns as float64, so it could not see
it. Verified on pandas 2.1.4: distinct values -> category, duplicate -> float64.

p90 is now cast to float after the map. A new test builds the event frame from
the real ingested slice with the downcast applied and distinct p90 values; it
failed with the same TypeError before the cast.

Running test_ml.py end to end also exposed a second a2f0e84 leftover: the
threshold leak guard recomputed percentiles on a float64 frame while the
pipeline publishes them from float32, and pressure differed in the sixth digit
(2.50132446 vs 2.50133672). The guard now recomputes on the downcast frame, as
its own comment says it should. _downcast_paired's docstring claimed values
were preserved exactly; it now says what float32 actually does, and warns about
the categorical map.

inference._per_variable_table maps `variable` the same way. It is safe today
because read_dataset returns object columns, and becomes unsafe the day serving
reads categoricals.

The run that failed: peak RSS 11,068,325,888 B, peak footprint 42.9 GB,
75,208,857 paired rows.

    tests/test_ml.py  14 passed

+++ backend/app/features/pivot.py
@@ -17,7 +17,10 @@ def build_event_frame(
 ) -> pd.DataFrame:
     df = paired.copy()
     df["pred_err"] = pred_err.reindex(df.index).to_numpy()
-    df["p90"] = df["variable"].map(p90_error)
+    # `variable` is categorical in the retrain's paired frame. Series.map on a categorical
+    # returns a categorical when every category maps to a distinct value - true for a
+    # real year, where no two variables share a p90 - and a categorical cannot be divided.
+    df["p90"] = df["variable"].map(p90_error).astype(float)
     df["conf"] = (1.0 - df["pred_err"] / df["p90"]).clip(CONF_FLOOR, 1.0)
 
     em = (df.groupby(EVENT_KEYS + ["variable"], observed=True)
